+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -125,7 +125,7 @@ import org.wordpress.android.ui.reader.views.ReaderSiteHeaderView;
 import org.wordpress.android.ui.reader.views.ReaderTagHeaderView;
 import org.wordpress.android.ui.reader.views.ReaderWebView;
 import org.wordpress.android.ui.sitecreation.NewSiteCreationActivity;
-import org.wordpress.android.ui.sitecreation.NewSiteCreationCategoryFragment;
+import org.wordpress.android.ui.sitecreation.segments.NewSiteCreationSegmentsFragment;
 import org.wordpress.android.ui.sitecreation.NewSiteCreationDomainAdapter;
 import org.wordpress.android.ui.sitecreation.NewSiteCreationDomainFragment;
 import org.wordpress.android.ui.sitecreation.NewSiteCreationDomainLoaderFragment;
@@ -225,7 +225,7 @@ public interface AppComponent extends AndroidInjector<WordPress> {
 
     void inject(NewSiteCreationActivity object);
 
-    void inject(NewSiteCreationCategoryFragment object);
+    void inject(NewSiteCreationSegmentsFragment object);
 
     void inject(NewSiteCreationThemeFragment object);
