+++ WordPress/src/main/res/layout/new_site_creation_segments_screen.xml
